participant-integration-api: Move configuration code into a new package.
