+++ ledger/participant-integration-api/src/test/suite/scala/platform/apiserver/configuration/LedgerConfigProvisionerSpec.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.apiserver.services
+package com.daml.platform.apiserver.configuration
 
 import java.time.{Duration, Instant}
 import java.util.concurrent.atomic.AtomicReference
